Refactor Calypso boot to avoid code duplication in Jetpack Cloud boot
