+++ client/boot/app.js
@@ -2,46 +2,15 @@
 import './polyfills';
 
 /**
- * External dependencies
- */
-import debugFactory from 'debug';
-import page from 'page';
-
-/**
  * Internal dependencies
  */
-import { configureReduxStore, setupMiddlewares, utils } from './common';
-import { setupLocale } from './locale';
-import { createReduxStore } from 'state';
-import initialReducer from 'state/reducer';
-import { getInitialState, persistOnChange, loadAllState } from 'state/initial-state';
-import detectHistoryNavigation from 'lib/detect-history-navigation';
-import userFactory from 'lib/user';
+import { bootApp } from './common';
 
 /**
  * Style dependencies
  */
 import 'assets/stylesheets/style.scss';
 
-const debug = debugFactory( 'calypso' );
-
-const boot = currentUser => {
-	debug( "Starting Calypso. Let's do this." );
-
-	utils();
-	loadAllState().then( () => {
-		const initialState = getInitialState( initialReducer );
-		const reduxStore = createReduxStore( initialState, initialReducer );
-		persistOnChange( reduxStore );
-		setupLocale( currentUser.get(), reduxStore );
-		configureReduxStore( currentUser, reduxStore );
-		setupMiddlewares( currentUser, reduxStore );
-		detectHistoryNavigation.start();
-		page.start( { decodeURLComponents: false } );
-	} );
-};
-
 window.AppBoot = () => {
-	const user = userFactory();
-	user.initialize().then( () => boot( user ) );
+	bootApp( 'Calypso' );
 };
